Module Sihl_email

module Sig = Sihl.Email.Sig

module Template : sig ... end

module Make : sig ... end

module MakeDelayed : functor (EmailService : Sig.SERVICE) -> functor (DbService : Sihl.Database.Sig.SERVICE) -> functor (QueueService : Sihl.Queue.Sig.SERVICE) -> Sig.SERVICE

Use this functor to create an email service that sends emails using the job queue. This is useful if you need to answer a request quickly while sending the email in the background