Module Sihl_email.MakeDelayed

Use this functor to create an email service that sends emails using the job queue. This is useful if you need to answer a request quickly while sending the email in the background

Parameters

EmailService : Sig.SERVICE

QueueService : Sihl.Queue.Sig.SERVICE

Signature

include Sig.Core.Container.Service.Sig

val lifecycle : Sihl_core.Container.Lifecycle.t

module Template : Sig.TEMPLATE_SERVICE

A template service to manage email templates.

val send : Sig.Core.Ctx.t -> Sihl_email_core__.Model.t -> unit Lwt.t

Send email.

val bulk_send : Sig.Core.Ctx.t -> Sihl_email_core__.Model.t list -> unit Lwt.t

Send multiple emails. If sending of one of them fails, the function fails.

val configure : Sig.Core.Configuration.data -> Sig.Core.Container.Service.t