Parameter MakeQueued.2-QueueService

include Sihl.Queue.Sig.Core.Container.Service.Sig

val lifecycle : Sihl_core.Container.Lifecycle.t

val dispatch : job:'a Sihl.Queue.Sig.Job.t -> ?⁠delay:Sihl.Queue.Sig.Utils.Time.duration -> 'a -> unit Lwt.t

Queue a job for processing. Use delay to run the initially job after a certain amount of time.

val register_jobs : jobs:'a Sihl.Queue.Sig.Job.t list -> unit Lwt.t

Register jobs that can be dispatched.

Only registered jobs can be dispatched. Dispatching a job that was not registered does nothing.

val register : ?⁠jobs:'a Sihl.Queue.Sig.Job.t list -> unit -> Sihl.Queue.Sig.Core.Container.Service.t