Parameter MakeMariaDb.1-MigrationService

include Sihl_core.Container.Service.Sig

val lifecycle : Sihl_core.Container.Lifecycle.t

val register_migration : Sihl_contract.Migration.t -> unit

register_migration migration registers a migration migration with the migration service so it can be executed with `run_all`.

val register_migrations : Sihl_contract.Migration.t list -> unit

register_migrations migrations registers migrations migrations with the migration service so it can be executed with `run_all`.

val execute : Sihl_contract.Migration.t list -> unit Lwt.t

execute migrations runs all migrations migrations.

val run_all : unit -> unit Lwt.t

run_all () runs all migrations that have been registered.

val register : ?⁠migrations:Sihl_contract.Migration.t list -> unit -> Sihl_core.Container.Service.t