Module Sihl_email

module Core = Sihl_core

module Database = Sihl_database

module Migration = Sihl_migration

module Repository = Sihl_repository

module Utils = Sihl_utils

module Queue = Sihl_queue_core

module Sig = Sihl.Email.Sig

module Template : sig ... end

module Make : sig ... end

module MakeDelayed : functor (EmailService : Sig.SERVICE) -> functor (QueueService : Sihl.Queue.Sig.SERVICE) -> Sig.SERVICE

Use this functor to create an email service that sends emails using the job queue. This is useful if you need to answer a request quickly while sending the email in the background